+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add props `is_hoverable` and `href` to `BNavbarDropdownItem`.
 - Add prop `id` to `BTitle`.
 - Add installation guide for Leptos CSR on the website.
+- Add customization with Sass section on the website.
 
 ## [0.0.0-alpha.13] - 2024-04-19